requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
discord.py==1.7.3
idna==3.2
multidict==5.1.0
typing-extensions==4.2.0.0
supabase==0.7.1
python-dotenv==0.21.0
psycopg2==2.9.5